Module rusoto_core::reactor

Default implementations for DispatchSignedRequest and ProvideAwsCredentials backed by an implicit event loop.

The purpose of this module is to enable the use of Rusoto without requiring any knowledge about tokio.

It does this by providing two simple structs, CredentialsProvider and RequestDispatcher, which can be plugged into any AWS client exposed by Rusoto.

These implementations are backed by an implicit event loop running on a background thread, so no manual event loop setup is required.

The implicit event loop is crated lazily on first use of these implementations, meaning that advanced users who wish to interface more directly with tokio can do so by using the otherwise provided implementations without causing any extra background threads or event loops to be spawned.

Structs

CredentialsProvider
A credentials provider backed by an implicit event loop.

CredentialsProviderFuture
Future returned from CredentialsProvider.

RequestDispatcher
A request dispatcher backed by an implicit event loop.

RequestDispatcherFuture
Future returned from RequestDispatcher.