Enum rusoto_core::region::Region

pub enum Region {
    ApEast1,
    ApNortheast1,
    ApNortheast2,
    ApSouth1,
    ApSoutheast1,
    ApSoutheast2,
    CaCentral1,
    EuCentral1,
    EuWest1,
    EuWest2,
    EuWest3,
    EuNorth1,
    SaEast1,
    UsEast1,
    UsEast2,
    UsWest1,
    UsWest2,
    UsGovEast1,
    UsGovWest1,
    CnNorth1,
    CnNorthwest1,
    Custom {
        name: String,
        endpoint: String,
    },
}

An AWS region.

Default

Region implements the Default trait. Calling Region::default() will attempt to read the AWS_DEFAULT_REGION or AWS_REGION environment variable. If it is malformed, it will fall back to Region::UsEast1. If it is not present it will fallback on the value associated with the current profile in ~/.aws/config or the file specified by the AWS_CONFIG_FILE environment variable. If that is malformed of absent it will fall back on Region::UsEast1

AWS-compatible services

Region::Custom can be used to connect to AWS-compatible services such as DynamoDB Local or Ceph.

    Region::Custom {
        name: "eu-east-3".to_owned(),
        endpoint: "http://localhost:8000".to_owned(),
    };

Caveats

CnNorth1 is currently untested due to Rusoto maintainers not having access to AWS China.

Variants
ApEast1

Region that covers the Eastern part of Asia Pacific

ApNortheast1

Region that covers the North-Eastern part of Asia Pacific

ApNortheast2

Region that covers the North-Eastern part of Asia Pacific

ApSouth1

Region that covers the Southern part of Asia Pacific

ApSoutheast1

Region that covers the South-Eastern part of Asia Pacific

ApSoutheast2

Region that covers the South-Eastern part of Asia Pacific

CaCentral1

Region that covers Canada

EuCentral1

Region that covers Central Europe

EuWest1

Region that covers Western Europe

EuWest2

Region that covers Western Europe

EuWest3

Region that covers Western Europe

EuNorth1

Region that covers Northern Europe

SaEast1

Region that covers South America

UsEast1

Region that covers the Eastern part of the United States

UsEast2

Region that covers the Eastern part of the United States

UsWest1

Region that covers the Western part of the United States

UsWest2

Region that covers the Western part of the United States

UsGovEast1

Region that covers the Eastern part of the United States for the US Government

UsGovWest1

Region that covers the Western part of the United States for the US Government

CnNorth1

Region that covers China

CnNorthwest1

Region that covers North-Western part of China

Custom

Specifies a custom region, such as a local Ceph target

Fields of Custom

name: String

Name of the endpoint (e.g. "eu-east-2").

endpoint: String

Endpoint to be used. For instance, "https://s3.my-provider.net" or just "s3.my-provider.net" (default scheme is https).

Methods

impl Region

pub fn name(&self) -> &str

Name of the region

    assert_eq!(Region::CaCentral1.name(), "ca-central-1");
    assert_eq!(
        Region::Custom { name: "eu-east-3".to_owned(), endpoint: "s3.net".to_owned() }.name(),
        "eu-east-3"
    );