Module rusoto_core::request

AWS API requests.

Wraps the hyper library to send PUT, POST, DELETE and GET requests.

Structs

BufferedHttpResponse
Stores the buffered response from a HTTP request.

HttpClient
Http client for use with AWS services.

HttpConfig
Configuration options for the HTTP Client

HttpDispatchError
An error produced when sending the request, such as a timeout error.

HttpResponse
Stores the response from a HTTP request.

TlsError
An error produced when the user has an invalid TLS client

Traits

DispatchSignedRequest
Trait for implementing HTTP Request/Response

Type Definitions

DispatchSignedRequestFuture
Type returned from dispatch for a DispatchSignedRequest implementor