Module rusoto_core::signature

AWS API request signatures.

Follows AWS Signature 4 algorithm.

If needed, the request will be re-issued to a temporary redirect endpoint. This can happen with newly created S3 buckets not in us-standard/us-east-1.

Structs

SignedRequest
A data structure for all the elements of an HTTP request that are involved in the Amazon Signature Version 4 signing process

StrictEncodeSet
This struct is used to maintain the strict URI encoding standard as proposed by RFC 3986

StrictPathEncodeSet
This struct is used to maintain the URI path encoding

Enums

SignedRequestPayload
Possible payloads included in a SignedRequest.

Statics

EMPTY_SHA256_HASH
Payload string to use for signed empty payload

UNSIGNED_PAYLOAD
Payload string to use for unsigned payload

Functions

string_to_sign
Mark string as AWS4-HMAC-SHA256 hashed