Struct rusoto_core::RusotoFuture

pub struct RusotoFuture<T, E> { /* fields omitted */ }

Future that is returned from all rusoto service APIs.

Mocking

To mock service traits, you can use [RusotoFuture::from_future] to create [RusotoFuture] instance. You can also use the [From] implementation on the [Result] value.

use std::time::{Duration, Instant};
use futures::prelude::*;
use rusoto_core::RusotoFuture;
use rusoto_s3::*;
use tokio_timer::Delay;

pub struct S3Mock;

impl S3 for S3Mock {
    fn abort_multipart_upload(
        &self,
        _input: AbortMultipartUploadRequest,
    ) -> RusotoFuture<AbortMultipartUploadOutput, AbortMultipartUploadError> {
        unimplemented!();
    }

    // ...

    fn put_object(&self, input: PutObjectRequest) -> RusotoFuture<PutObjectOutput, PutObjectError> {
        if input.bucket == "foo" {
            let deadline = Instant::now() + Duration::from_secs(3);
            let output = PutObjectOutput {
                ..Default::default()
            };
            RusotoFuture::from_future(
                Delay::new(deadline)
                    .map_err(|e| PutObjectError::Unknown(e.to_string()))
                    .map(|_| output)
            )
        } else {
            Err(PutObjectError::Validation("Invalid bucket".to_string())).into()
        }
    }

    // ...
}

Methods

impl<T, E> RusotoFuture<T, E>

pub fn with_timeout(self, timeout: Duration) -> Self

Set the timeout on the future to the provided duration.

Unlike set_timeout this method can be easily chained:

let future = s3.list_buckets()
    .with_timeout(Duration::from_secs(10));

This is only guaranteed to take effect when called before the future is polled for the first time.

pub fn set_timeout(&mut self, timeout: Duration)

Set the timeout on the future to the provided duration.

This is only guaranteed to take effect when called before the future is polled for the first time.

pub fn clear_timeout(&mut self)

Clear the timeout on the future.

This is only guaranteed to take effect when called before the future is polled for the first time.

pub fn sync(self) -> RusotoResult<T, E> where
T: Send + 'static,
E: Send + 'static,

Blocks the current thread until the future has resolved.

This is meant to provide a simple way for non-async consumers to work with rusoto.

pub fn from_future<F>(fut: F) -> Self where
F: IntoFuture<Item = T, Error = RusotoError<E>>,
F::Future: Send + 'static,

Wraps the provided future, mainly to mock the service response.

Caution

This is not intended to be used outside of the test case. In production, [Box] is recommended.