Crate rusoto_core

Rusoto is an AWS SDK for Rust. A high level overview is available in README.md at https://github.com/rusoto/rusoto.

Re-exports

pub extern crate rusoto_credential as credential;

pub use crate::region::Region;

pub use crate::request::DispatchSignedRequest;

pub use crate::request::HttpClient;

pub use crate::request::HttpConfig;

pub use crate::request::HttpDispatchError;

Modules

param
Parameters for talking to query-based AWS services.

region
AWS Regions and helper functions.

request
AWS API requests.

signature
AWS API request signatures.

Structs

ByteStream
Stream of bytes.

CredentialsError
Represents an Error that has occured during the fetching Credentials Phase.

DefaultCredentialsProvider
Wraps a ChainProvider in an AutoRefreshingProvider.

RusotoFuture
Future that is returned from all rusoto service APIs.

Enums

RusotoError
Generic error type returned by all rusoto requests.

Traits

ProvideAwsCredentials
A trait for types that produce AwsCredentials.

Type Definitions

RusotoResult
Result carrying a generic RusotoError.