Struct rusoto_core::request::HttpClient

pub struct HttpClient<C = HttpsConnector<HttpConnector>> { /* fields omitted */ }

Http client for use with AWS services.

Implementations

impl HttpClient

pub fn new() -> Result<Self, TlsError>

Create a tls-enabled http client.

pub fn new_with_config(config: HttpConfig) -> Result<Self, TlsError>

Create a tls-enabled http client.

pub fn local_agent(&mut self, local_agent: String)

An alias for [local_agent_prepend] for backwards compatibility

pub fn local_agent_prepend(&mut self, local_agent: String)

Sets a local agent that is prepended to the default HTTP User-Agent used by Rusoto.

pub fn local_agent_append(&mut self, local_agent: String)

Sets a local agent that is appended to the default HTTP User-Agent used by Rusoto.

impl<C> HttpClient<C> where
C: Connect + Clone + Send + Sync,

pub fn from_connector(connector: C) -> Self

Allows for a custom connector to be used with the HttpClient

pub fn from_connector_with_config(connector: C, config: HttpConfig) -> Self

Allows for a custom connector to be used with the HttpClient with extra configuration options

pub fn from_builder(builder: HyperBuilder, connector: C) -> Self

Alows for a custom builder and connector to be used with the HttpClient