Struct rusoto_core::BaseAutoRefreshingProvider

pub struct BaseAutoRefreshingProvider<P, T> where
    P: 'static + ProvideAwsCredentials,  { /* fields omitted */ }

Wrapper for ProvideAwsCredentials that caches the credentials returned by the wrapped provider. Each time the credentials are accessed, they are checked to see if they have expired, in which case they are retrieved from the wrapped provider again.

In order to access the wrapped provider, for instance to set a timeout, the get_ref and get_mut methods can be used.

Methods

impl<P, T> BaseAutoRefreshingProvider<P, T> where
P: 'static + ProvideAwsCredentials,

Important traits for &'a mut R
impl<'a, R> Read for &'a mut R where
R: Read + ?Sized, impl<'a, W> Write for &'a mut W where
W: Write + ?Sized, impl<'a, I> Iterator for &'a mut I where
I: Iterator + ?Sized, type Item = <I as Iterator>::Item;

pub fn get_ref(&self) -> &P

Get a shared reference to the wrapped provider.

Important traits for &'a mut R
impl<'a, R> Read for &'a mut R where
R: Read + ?Sized, impl<'a, W> Write for &'a mut W where
W: Write + ?Sized, impl<'a, I> Iterator for &'a mut I where
I: Iterator + ?Sized, type Item = <I as Iterator>::Item;

pub fn get_mut(&mut self) -> &mut P

Get a mutable reference to the wrapped provider.

This can be used to call set_timeout on the wrapped provider.

impl<P> BaseAutoRefreshingProvider<P, Mutex<Shared<<P as ProvideAwsCredentials>::Future>>> where
P: 'static + ProvideAwsCredentials,

pub fn with_mutex(
provider: P
) -> Result<BaseAutoRefreshingProvider<P, Mutex<Shared<<P as ProvideAwsCredentials>::Future>>>, CredentialsError>

Grab a RefreshingProvider that locks it's credentials with a Mutex so it's thread safe.

impl<P> BaseAutoRefreshingProvider<P, RefCell<Shared<<P as ProvideAwsCredentials>::Future>>> where
P: 'static + ProvideAwsCredentials,

pub fn with_refcell(
provider: P
) -> Result<BaseAutoRefreshingProvider<P, RefCell<Shared<<P as ProvideAwsCredentials>::Future>>>, CredentialsError>

Grab a provider that locks it's credentials with a RefCell. If you're looking for Thread Safety, take a look at AutoRefreshingProviderSync.

impl BaseAutoRefreshingProvider<ChainProvider, RefCell<Shared<<ChainProvider as ProvideAwsCredentials>::Future>>>

pub fn new(
handle: &Handle
) -> Result<BaseAutoRefreshingProvider<ChainProvider, RefCell<Shared<<ChainProvider as ProvideAwsCredentials>::Future>>>, CredentialsError>

Creates a new DefaultCredentials Provider. If you're looking for Thread Safety look at DefaultCredentialsProviderSync.

impl BaseAutoRefreshingProvider<ChainProvider, Mutex<Shared<<ChainProvider as ProvideAwsCredentials>::Future>>>

pub fn new(
handle: &Handle
) -> Result<BaseAutoRefreshingProvider<ChainProvider, Mutex<Shared<<ChainProvider as ProvideAwsCredentials>::Future>>>, CredentialsError>

Creates a new Thread Safe Default Credentials Provider.