Struct rusoto_core::BaseAutoRefreshingProvider

pub struct BaseAutoRefreshingProvider<P, T> where
    P: 'static + ProvideAwsCredentials,  { /* fields omitted */ }

Wrapper for ProvideAwsCredentials that caches the credentials returned by the wrapped provider. Each time the credentials are accessed, they are checked to see if they have expired, in which case they are retrieved from the wrapped provider again.

Methods

impl<P> BaseAutoRefreshingProvider<P, Mutex<Shared<<P as ProvideAwsCredentials>::Future>>> where
P: 'static + ProvideAwsCredentials,

pub fn with_mutex(
provider: P
) -> Result<BaseAutoRefreshingProvider<P, Mutex<Shared<<P as ProvideAwsCredentials>::Future>>>, CredentialsError>

Grab a RefreshingProvider that locks it's credentials with a Mutex so it's thread safe.

impl<P> BaseAutoRefreshingProvider<P, RefCell<Shared<<P as ProvideAwsCredentials>::Future>>> where
P: 'static + ProvideAwsCredentials,

pub fn with_refcell(
provider: P
) -> Result<BaseAutoRefreshingProvider<P, RefCell<Shared<<P as ProvideAwsCredentials>::Future>>>, CredentialsError>

Grab a provider that locks it's credentials with a RefCell. If you're looking for Thread Safety, take a look at AutoRefreshingProviderSync.

impl BaseAutoRefreshingProvider<ChainProvider, RefCell<Shared<<ChainProvider as ProvideAwsCredentials>::Future>>>

pub fn new(
handle: &Handle
) -> Result<BaseAutoRefreshingProvider<ChainProvider, RefCell<Shared<<ChainProvider as ProvideAwsCredentials>::Future>>>, CredentialsError>

Creates a new DefaultCredentials Provider. If you're looking for Thread Safety look at DefaultCredentialsProviderSync.

impl BaseAutoRefreshingProvider<ChainProvider, Mutex<Shared<<ChainProvider as ProvideAwsCredentials>::Future>>>

pub fn new(
handle: &Handle
) -> Result<BaseAutoRefreshingProvider<ChainProvider, Mutex<Shared<<ChainProvider as ProvideAwsCredentials>::Future>>>, CredentialsError>

Creates a new Thread Safe Default Credentials Provider.